+++ fixtures/thirdparty/src/styles.css.ts
@@ -1,0 +1,18 @@
+import { style, createVar } from '@vanilla-extract/css';
+
+export {
+  depBlock,
+  depColor,
+  depdepBlock,
+  depdepColor,
+  // @ts-expect-error no types
+} from '@fixtures/thirdparty-dep';
+
+const color = createVar();
+
+export const block = style({
+  vars: {
+    [color]: 'red',
+  },
+  backgroundColor: color,
+});