+++ fixtures/thirdparty/thirdparty-dep/thirdparty-dep-dep/styles.css.mjs
@@ -1,0 +1,10 @@
+import { style, createVar } from '@vanilla-extract/css';
+
+export const depdepColor = createVar();
+
+export const depdepBlock = style({
+  vars: {
+    [depdepColor]: 'blue',
+  },
+  backgroundColor: depdepColor,
+});